puddle/requirements.txt
amqp==5.3.1
asgiref==3.9.1
asttokens==3.0.0
billiard==4.2.1
celery==5.5.3
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coverage==7.10.3
cron-descriptor==1.4.5
decorator==5.2.1
dj-database-url==3.0.1
Django==5.2.4
django-celery-beat==2.8.1
django-debug-toolbar==6.0.0
django-timezone-field==7.1
executing==2.2.0
flower==2.0.1
humanize==4.12.3
iniconfig==2.1.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
kombu==5.5.4
matplotlib-inline==0.1.7
packaging==25.0
parso==0.8.4
pillow==11.3.0
pluggy==1.6.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
psycopg==3.2.9
psycopg2==2.9.10
pure_eval==0.2.3
Pygments==2.19.2
pytest==8.4.1
pytest-cov==6.2.1
pytest-django==4.11.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
redis==6.4.0
reportlab==4.4.3
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
gunicorn==23.0.0

# REST API
djangorestframework==3.15.2
drf-spectacular==0.27.2
django-cors-headers==4.6.0
django-filter==24.3


